added vfs and multiblock support, though it panics almost right away lol

+++ user/libc/include/sys/sysfn.h
@@ -38,5 +38,5 @@ int reboot();
 int poweroff();
 void sleep(int secs);
 int termctl(int code, int arg);
-int open(char* path, u32 flags, int mode);
+int open(char* path, int flags, u16 mode);
 int close(int fd);
